+++ apps/modernization-ui/src/components/NoData/NoData.scss
@@ -1,6 +1,7 @@
 .no-data {
     $no-data-color: #757575;
-
+    min-width: 8ch;
+    display: inline-block;
     color: $no-data-color;
     margin: 0;
     font-style: italic;
