--- apps/modernization-ui/src/design-system/table/height-clamp.module.scss
@@ -1,14 +1,0 @@
-.constrained {
-    max-height: 42px;
-    overflow: hidden;
-}
-.expander {
-    max-height: unset;
-}
-.button {
-  padding: 0 !important;
-  margin-left: auto !important;
-  width: 9ch !important;
-  float: right;
-  text-decoration: unset !important;
-}
