+++ apps/modernization-ui/src/design-system/table/height-constrained.module.scss
@@ -1,0 +1,14 @@
+.bounded {
+    max-height: 42px;
+    overflow: hidden;
+}
+.unbounded {
+    max-height: unset;
+}
+.button {
+    padding: 0 !important;
+    margin-left: auto !important;
+    width: 9ch !important;
+    float: right;
+    text-decoration: unset !important;
+}
